Run on Mac
Currently this requires installing MacPorts (https://www.macports.org/), then `sudo port install csfml`. This is typically a reasonable requirement, but in this case I’ll continue to search for something easier.

The reason for this is the SFML.Net NuGet package doesn’t include binaries for OSX, which is really annoying. The details of Mac’s security system prevent us from simply including unsigned binaries in the project.